prediction-model/requirements.txt
pySerial
pandas 
numpy
matplotlib
tensorflow